+++ src/Field.php
@@ -685,6 +685,8 @@ class Field extends HtmlField implements ElementContainerFieldInterface
                     'imageTextAlternative',
                 ],
             ],
+            'assetSources' => $this->_assetSources(),
+            'assetSelectionCriteria' => $this->_assetSelectionCriteria(),
             'linkOptions' => $this->_linkOptions($element),
             'table' => [
                 'contentToolbar' => [
@@ -1264,9 +1266,9 @@ JS,
     {
         $linkOptions = [];
 
-        $sectionSources = $this->_sectionSources($element);
+        $sectionSources = $this->_entrySources($element);
         $categorySources = $this->_categorySources($element);
-        $volumeSources = $this->_volumeSources();
+        $volumeSources = $this->_assetSources(true);
 
         if (!empty($sectionSources)) {
             $linkOptions[] = [
@@ -1289,16 +1291,12 @@ JS,
         }
 
         if (!empty($volumeSources)) {
-            $criteria = [];
-            if ($this->showUnpermittedFiles) {
-                $criteria['uploaderId'] = null;
-            }
             $linkOptions[] = [
                 'label' => Craft::t('ckeditor', 'Link to an asset'),
                 'elementType' => Asset::class,
                 'refHandle' => Asset::refHandle(),
                 'sources' => $volumeSources,
-                'criteria' => $criteria,
+                'criteria' => $this->_assetSelectionCriteria(),
             ];
         }
 
@@ -1322,13 +1320,13 @@ JS,
     }
 
     /**
-     * Returns the available section sources.
+     * Returns the available entry sources.
      *
      * @param ElementInterface|null $element The element the field is associated with, if there is one
      * @param bool $showSingles Whether to include Singles in the available sources
      * @return array
      */
-    private function _sectionSources(?ElementInterface $element, bool $showSingles = false): array
+    private function _entrySources(?ElementInterface $element, bool $showSingles = false): array
     {
         $sources = [];
         $sections = Craft::$app->getEntries()->getAllSections();
@@ -1396,11 +1394,12 @@ JS,
     }
 
     /**
-     * Returns the available volume sources.
+     * Returns the available asset sources.
      *
+     * @param bool $withUrlsOnly Whether to only return volumes that have filesystems that have public URLs
      * @return string[]
      */
-    private function _volumeSources(): array
+    private function _assetSources(bool $withUrlsOnly = false): array
     {
         if (!$this->availableVolumes) {
             return [];
@@ -1417,6 +1416,11 @@ JS,
             $volumes = $volumes->filter(fn(Volume $volume) => $userService->checkPermission("viewAssets:$volume->uid"));
         }
 
+        if ($withUrlsOnly) {
+            // only allow volumes that belong to FS that have public URLs
+            $volumes = $volumes->filter(fn(Volume $volume) => $volume->getFs()->hasUrls);
+        }
+
         $sources = $volumes
             ->map(fn(Volume $volume) => "volume:$volume->uid")
             ->values()
@@ -1429,6 +1433,20 @@ JS,
         }
 
         return $sources;
+    }
+
+    /**
+     * Returns the asset selection criteria.
+     *
+     * @return array
+     */
+    private function _assetSelectionCriteria(): array
+    {
+        $criteria = [];
+        if ($this->showUnpermittedFiles) {
+            $criteria['uploaderId'] = null;
+        }
+        return $criteria;
     }
 
     /**